+++ lookup_extensions/sql.py
@@ -1,20 +1,15 @@
-from django.core.exceptions import FieldError
-from django.db.models import Exists
 from django.db.models.constants import LOOKUP_SEP
 from django.db.models.sql import Query as DjangoQuery
 
 
 class ExtendedQueryMixin(object):
     SUBQUERY_LOOKUPS = (
-        'exists',
-        'neexists',
+        'complement',
     )
 
     def build_lookup(self, lookups, lhs, rhs):
         if lookups and lookups[-1] in self.SUBQUERY_LOOKUPS:
-            if not isinstance(rhs, Exists):
-                raise FieldError("Value is not Subquery instance.")
-            return super(ExtendedQueryMixin, self).build_lookup(['exact'], rhs, not lookups[-1].startswith('ne'))
+            return super(ExtendedQueryMixin, self).build_lookup(lookups[-1:], rhs, rhs)
         return super(ExtendedQueryMixin, self).build_lookup(lookups, lhs, rhs)
 
     def prepare_lookup_value(self, value, lookups, can_reuse, allow_joins=True):