+++ HISTORY.rst
@@ -7,6 +7,7 @@ History
 
 - Add support for Django 2.1
 - Add support for Python 3.7
+- Migrate to `complement` from `exists` and `neexists`
 
 0.1.0
 -----